Tempo was integer and become double : 120.2 is ok  now

+++ chordv/src/dialognewsong.h
@@ -35,7 +35,7 @@ signals:
     /// \param time1 : default 4 :  4 / 4 the second 4 is time2
     /// \param time2 : see time1
     ///
-    void InsertSong(QString title, QString subtitle, bool compress, int columns, int tempo, int time1 , int time2);
+    void InsertSong(QString title, QString subtitle, bool compress, int columns, double tempo, int time1 , int time2);
 private slots:
     ///
     /// \brief ActiveInsertButton set or unset insert button : enable only if title is not empty
